Move nexus object to correct directory.. fix imports,
